Google Search Form Intext
In this file you have about 47 links to select and you can even use emojis!

Instead of entering several times intext: Operator and/or emoji and/or entering links - this can be annoying. This file tries to save you time!

Hint: You can use this with safari and firefox browser only!